Debugged last commit

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -11,7 +11,7 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-int l1 = strlen(dest);
+size_t l1 = strlen(dest);
 int i;
 
 for (i = 0; i < n; i++)
